requirements.txt
streamlit
plotly
pymongo
python-dotenv
